requirements.txt
Flask==2.3.3
twilio==8.10.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
openai==1.12.0
Flask-SQLAlchemy==3.0.5
